+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
   - Change `status_code` to `status` for SDX logging
+  - Change `Calling service` and `Returned from service` to add context
 
 ### 1.4.0 2017-02-16
   - Add change log